Logging to db is working but with multiple entries.

+++ src/backend/apps/base/models/test_result.py
@@ -5,7 +5,7 @@ from ..models import TestCase
 class TestResult(models.Model):
     test_case = models.ForeignKey(TestCase, related_name='test_result')
     field = models.CharField(max_length=32, blank=True, null=True)
-    value = models.IntegerField(blank=True, null=True)
+    value = models.FloatField(blank=True, null=True)
     timestamp = models.DateTimeField()
 
     def __str__(self):
